Optimize DIC engine: Migrate to 32-bit float and NEON SIMD
- Migrated core engine from 64-bit double to 32-bit float (Types.h, OptimizationEngine, SubsetPrecomputer, StrainCalculator)
- Reduced RAM bandwidth requirements by 50%
- Upgraded ARM NEON intrinsics from float64x2_t to float32x4_t (4 flops/cycle)
- Unrolled bicubic interpolation to bypass loop overhead
- Replaced indirect array lookups with contiguous memory prefetching in SubsetPrecomputer
- Maintained perfect matrix conditioning via subset ZNSSD normalization
- Benchmark: 41,685 points processed in ~2.7s (down from 3.9s)

+++ app/src/main/cpp/core/OptimizationEngine.h
@@ -10,20 +10,25 @@ namespace IndicVision {
 
     class OptimizationEngine {
     public:
-        // Profiling trackers
+        // Profiling trackers (kept as double for timing precision)
         double time_icgn_ms = 0.0;
         double time_simplex_ms = 0.0;
         int count_icgn = 0;
         int count_simplex = 0;
-        std::vector<double> eval_buffer;
+
+        // 🚀 Converted evaluation buffer to float
+        std::vector<float> eval_buffer;
 
         AnalysisResult calculate_deformation(const SubsetData& subset, const Image& def_img, scalar_t guess_u, scalar_t guess_v, InitializationMode init_mode);
 
     private:
         void estimate_initial_guess(const SubsetData& subset, const Image& def_img, scalar_t& best_u, scalar_t& best_v);
-        AnalysisResult solve_icgn(const SubsetData& subset, const Image& def_img, double init_u, double init_v);
+
+        // 🚀 Core solvers now use pure float math
+        AnalysisResult solve_icgn(const SubsetData& subset, const Image& def_img, float init_u, float init_v);
         AnalysisResult solve_simplex(const SubsetData& subset, const Image& def_img, AnalysisResult start, bool translation_only);
-        double evaluate_znssd(const SubsetData& subset, const Image& def_img, double u, double v, double ux, double uy, double vx, double vy, std::vector<double>& buffer);
+
+        float evaluate_znssd(const SubsetData& subset, const Image& def_img, float u, float v, float ux, float uy, float vx, float vy, std::vector<float>& buffer);
     };
 
 }
